pyproject.toml
[project]
name = "peri"
version = "0.0.1"
description = "tools to print to the PeriA6 bluetooth receipt printer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pillow",
  "pyobjc-core",
  "pyobjc-framework-Cocoa",
  "pyobjc-framework-CoreBluetooth",
  "pyobjc-framework-IOBluetooth",
  "tqdm",
]

[project.scripts]
peri = "peri:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
